+++ spyne/interface/_base.py
@@ -170,6 +170,8 @@ class Interface(object):
                         self.__test_type_name_validity(in_header)
                         in_header.resolve_namespace(in_header, self.get_tns())
                         classes.append(in_header)
+                        if in_header.get_namespace() != self.get_tns():
+                            self.imports[self.get_tns()].add(in_header.get_namespace())
 
                 if not (method.out_header is None):
                     if not isinstance(method.out_header, (list, tuple)):
@@ -179,6 +181,8 @@ class Interface(object):
                         self.__test_type_name_validity(out_header)
                         out_header.resolve_namespace(out_header, self.get_tns())
                         classes.append(out_header)
+                        if out_header.get_namespace() != self.get_tns():
+                            self.imports[self.get_tns()].add(out_header.get_namespace())
 
                 if method.faults is None:
                     method.faults = []
@@ -201,6 +205,9 @@ class Interface(object):
                 method.out_message.resolve_namespace(method.out_message,
                                                                  self.get_tns())
                 classes.append(method.out_message)
+
+                for p in method.patterns:
+                    p.endpoint = method.key
 
         classes.sort(key=lambda cls: (cls.get_namespace(), cls.get_type_name()))
         for c in classes: