+++ spyne/protocol/_base.py
@@ -176,6 +176,8 @@ class ProtocolBase(object):
     type = set()
     """Set that contains keywords about a protocol."""
 
+    default_binary_encoding = None
+
     def __init__(self, app=None, validator=None, mime_type=None, skip_depth=0, ignore_uncap=False):
         self.__app = None
         self.validator = None