+++ spyne/util/wsgi_wrapper.py
@@ -25,10 +25,14 @@ logger = logging.getLogger(__name__)
 
 from spyne.server.wsgi import WsgiApplication
 
+
 class WsgiMounter(object):
     """Simple mounter object for wsgi callables. Takes a dict where the keys are
     uri fragments and values are :class:`spyne.application.Application`
     instances.
+
+    :param mounts: dict of :class:`spyne.application.Application` instances
+    whose keys are url fragments.
     """
 
     @staticmethod
@@ -37,8 +41,8 @@ class WsgiMounter(object):
         return []
 
     def __init__(self, mounts=None):
-        self.mounts = mounts or { }
-        self.mounts = dict([(k, WsgiApplication(v)) for k,v in self.mounts.items()])
+        self.mounts = dict([(k, WsgiApplication(v)) for k,v in
+                                                        (mounts or {}).items()])
 
     def __call__(self, environ, start_response):
         path_info = environ.get('PATH_INFO', '')
@@ -52,8 +56,8 @@ class WsgiMounter(object):
 
         original_script_name = environ.get('SCRIPT_NAME', '')
 
-        environ['SCRIPT_NAME'] = "/" + original_script_name + script
-        environ['PATH_INFO'] = '/' + '/'.join(fragments[1:])
+        environ['SCRIPT_NAME'] = ''.join(('/', original_script_name, script))
+        environ['PATH_INFO'] = ''.join(('/', '/'.join(fragments[1:])))
 
         return app(environ, start_response)